fix: deleted print button on patient data page. Added add patient button to patients table
