+++ src/components/mainPage/PatientsTable/PatientsTable.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Patient, PatientData } from "../../../types/models/Patient";
 import { patientDataFields } from "../../../validation-schemas/PatientDataSchema";
+import CreateNewPatientButton from "../../shared/CreateNewPatientButton";
 
 interface PatientTableProps {
     patients: Patient[];
@@ -16,32 +17,35 @@ const PatientsTable: React.FC<PatientTableProps> = ({ patients }) => {
     }
 
     return (
-        <div className="overflow-x-auto mr-8">
-            <table className="min-w-full bg-white border">
-                <thead className="bg-gray-200">
-                    <tr>
-                        {Object.values(patientDataFields).map((value) => {
-                            return (
-                                <th key={value.placeholder} className="px-4 py-2 border">
-                                    {value.placeholder}
-                                </th>
-                            );
-                        })}
-                    </tr>
-                </thead>
-                <tbody>
-                    {patients.map((patient) => (
-                        <tr key={patient.id}>
-                            {Object.keys(patientDataFields).map((key) => (
-                                <td key={patient.id + key} className="px-4 py-2 border">
-                                    {String(patient[key as keyof PatientData])}
-                                </td>
-                            ))}
+        <>
+            <CreateNewPatientButton className="mr-8 mb-8" value="Cтворити нового паціента" />
+            <div className="overflow-x-auto mr-8">
+                <table className="min-w-full bg-white border">
+                    <thead className="bg-gray-200">
+                        <tr>
+                            {Object.values(patientDataFields).map((value) => {
+                                return (
+                                    <th key={value.placeholder} className="px-4 py-2 border">
+                                        {value.placeholder}
+                                    </th>
+                                );
+                            })}
                         </tr>
-                    ))}
-                </tbody>
-            </table>
-        </div>
+                    </thead>
+                    <tbody>
+                        {patients.map((patient) => (
+                            <tr key={patient.id}>
+                                {Object.keys(patientDataFields).map((key) => (
+                                    <td key={patient.id + key} className="px-4 py-2 border">
+                                        {String(patient[key as keyof PatientData])}
+                                    </td>
+                                ))}
+                            </tr>
+                        ))}
+                    </tbody>
+                </table>
+            </div>
+        </>
     );
 };
